+++ test/Test/Universum/Property.hs
@@ -1,4 +1,4 @@
-{-# LANGUAGE CPP #-}
+{-# LANGUAGE CPP          #-}
 {-# LANGUAGE TypeFamilies #-}
 
 module Test.Universum.Property
@@ -8,23 +8,31 @@ module Test.Universum.Property
 import Universum
 
 import Data.List (nub)
-import Hedgehog (Property, Gen, MonadGen, forAll, property, assert, (===))
+import Hedgehog (Gen, MonadGen, Property, assert, forAll, property, (===))
 #if MIN_VERSION_hedgehog(1,0,0)
 import Hedgehog (GenBase)
 #endif
-import Test.Tasty (testGroup, TestTree)
+import Test.Tasty (TestTree, testGroup)
 import Test.Tasty.Hedgehog
 
+import qualified Data.ByteString as B
+import qualified Data.ByteString.Lazy as LB
+import qualified Data.Text as T
+import qualified Data.Text.Lazy as LT
+import qualified Hedgehog.Gen as Gen
+import qualified Hedgehog.Range as Range
 import qualified Universum as U
-import qualified Hedgehog.Gen              as Gen
-import qualified Hedgehog.Range            as Range
-import qualified Data.ByteString           as B
-import qualified Data.ByteString.Lazy      as LB
-import qualified Data.Text                 as T
-import qualified Data.Text.Lazy            as LT
 
 hedgehogTestTree :: TestTree
-hedgehogTestTree = testGroup "Tests" [utfProps, listProps, boolMProps]
+hedgehogTestTree = testGroup "Tests" [stringProps, utfProps, listProps, boolMProps]
+
+stringProps :: TestTree
+stringProps = testGroup "String conversions"
+    [ testProperty "toString . toText = id" prop_StringToTextAndBack
+    , testProperty "`toString . toText` for UTF-16 surrogate"
+        prop_StringToTextAndBackSurrogate
+    , testProperty "toText . toString = id" prop_TextToStringAndBack
+    ]
 
 utfProps :: TestTree
 utfProps = testGroup "utf8 conversion property tests"
@@ -47,8 +55,14 @@ unicode' = do
 utf8String :: Gen U.String
 utf8String = Gen.string (Range.linear 0 10000) unicode'
 
+unicodeAllString :: Gen U.String
+unicodeAllString = Gen.string (Range.linear 0 10000) Gen.unicodeAll
+
 utf8Text :: Gen T.Text
 utf8Text = Gen.text (Range.linear 0 10000) unicode'
+
+unicodeAllText :: Gen T.Text
+unicodeAllText = Gen.text (Range.linear 0 10000) Gen.unicodeAll
 
 utf8Bytes :: Gen B.ByteString
 utf8Bytes = Gen.utf8 (Range.linear 0 10000) unicode'
@@ -57,6 +71,33 @@ utf8Bytes = Gen.utf8 (Range.linear 0 10000) unicode'
 -- > import qualified Data.ByteString.UTF8 as BU
 -- > BU.toString (BU.fromString "\65534") == "\65533"
 -- > True
+
+prop_StringToTextAndBack :: Property
+prop_StringToTextAndBack = property $ do
+  str <- forAll unicodeAllString
+  toString (toText str) === str
+
+-- | See comment to this function:
+-- <http://hackage.haskell.org/package/text-1.2.3.1/docs/src/Data.Text.Internal.html#safe>
+--
+-- While 'String' may contain surrogate UTF-16 code points, actually UTF-8
+-- doesn't allow them, as well as 'Text'. 'Data.Text.pack' replaces invalid
+-- characters with unicode replacement character, so by default
+-- @toString . toText@ is not identity.
+--
+-- However, we have a rewrite rule by which we /replace/ @toString . toText@
+-- occurrences with the identity function.
+prop_StringToTextAndBackSurrogate :: Property
+prop_StringToTextAndBackSurrogate = property $ do
+  -- Surrogate character like this one should remain intact
+  -- Without rewrite rule this string would be transformed to "\9435"
+  let str = "\xD800"
+  toString (toText str) === str
+
+prop_TextToStringAndBack :: Property
+prop_TextToStringAndBack = property $ do
+  txt <- forAll unicodeAllText
+  toText (toString txt) === txt
 
 prop_StringToBytes :: Property
 prop_StringToBytes = property $ do